fix port filtering and ntohs(port) display in debug output

+++ dht.h
@@ -33,7 +33,7 @@ dht_callback(void *closure, int event,
 
 extern FILE *dht_debug;
 extern FILE *port_filter;
-extern int blocked_port;
+extern unsigned short blocked_port;
 
 int dht_init(int s, int s6, const unsigned char *id, const unsigned char *v);
 int dht_insert_node(const unsigned char *id, struct sockaddr *sa, int salen);
